pyproject.toml
[tool]
[tool.poetry]
name = "xcat"
version = "1.2.0"
description = "A command line tool to automate the exploitation of blind XPath injection vulnerabilities"
license = "MIT"
classifiers = ["Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License"]
homepage = "https://github.com/orf/xcat"
authors = ["Tom Forbes <tom@tomforb.es>"]
readme = "README.md"
include = [
    "LICENSE",
    "README.md"
]

[tool.poetry.scripts]
xcat = "xcat.cli:cli"

[tool.poetry.dependencies]
python = ">=3.10"
aiodns = "*"
aiohttp = ">=3.9,<4"
platformdirs = "*"
click = ">=8.0"
colorama = "*"
prompt-toolkit = ">=3.0,<4"
xpath-expressions = ">=1.0,<2"

[tool.poetry.group.dev.dependencies]
bump2version = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
